backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
sse-starlette>=2.1.0
python-dotenv>=1.0.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.150.0
pydantic>=2.9.0
